[MOVE] State 코드 분리 #23
